+++ packages/xchain-litecoin/src/client.ts
@@ -255,11 +255,11 @@ class Client implements LitecoinClient, XChainClient {
         const rawTx = await sochain.getTx(this.nodeUrl, this.net, txItem.txid)
         const tx: Tx = {
           asset: AssetLTC,
-          from: rawTx.inputs.map((i: TxIO) => ({ from: i.address, amount: assetToBase(assetAmount(i.value, 8)) })),
+          from: rawTx.inputs.map((i: TxIO) => ({ from: i.address, amount: assetToBase(assetAmount(i.value, Utils.LTC_DECIMAL)) })),
           to: rawTx.outputs
             // ignore tx with type 'nulldata'
             .filter((i: TxIO) => i.type !== 'nulldata')
-            .map((i: TxIO) => ({ to: i.address, amount: assetToBase(assetAmount(i.value, 8)) })),
+            .map((i: TxIO) => ({ to: i.address, amount: assetToBase(assetAmount(i.value, Utils.LTC_DECIMAL)) })),
           date: new Date(rawTx.time * 1000),
           type: 'transfer',
           hash: rawTx.txid,
@@ -288,8 +288,8 @@ class Client implements LitecoinClient, XChainClient {
       const rawTx = await sochain.getTx(this.nodeUrl, this.net, txId)
       return {
         asset: AssetLTC,
-        from: rawTx.inputs.map((i) => ({ from: i.address, amount: assetToBase(assetAmount(i.value, 8)) })),
-        to: rawTx.outputs.map((i) => ({ to: i.address, amount: assetToBase(assetAmount(i.value, 8)) })),
+        from: rawTx.inputs.map((i) => ({ from: i.address, amount: assetToBase(assetAmount(i.value, Utils.LTC_DECIMAL)) })),
+        to: rawTx.outputs.map((i) => ({ to: i.address, amount: assetToBase(assetAmount(i.value, Utils.LTC_DECIMAL)) })),
         date: new Date(rawTx.time * 1000),
         type: 'transfer',
         hash: rawTx.txid,